+++ NEWS.md
@@ -1,3 +1,5 @@
+# VIM 6.1.1
+- ordFun as parameter of kNN to control the function applied to ordinal variables
 # VIM 6.0.2
 - error message in `hotdeck()` when ord_var and variable overlap
 - family argument of class 'family' now work in regressionImp